Tangle.coffee: Improve choices when growing
If we connect to another node, return it instead of the new node,
so that it becomes active if the user is extending the graph.

Only allow nodes in front of us to become connected. This will be extended
to try and be smarter about which of the nodes in front of us is the best
choice to connect to.